+++ storybook-example/package.json
@@ -16,9 +16,23 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "storybook dev -p 6006",
+    "storybook": "storybook dev -p 6006",
     "build": "storybook build",
+    "build:react-sign-in-with-esignet": "cd ../react-sign-in-with-esignet && npm run build-storybook -- -o ../storybook-example/storybook-static/react-sign-in-with-esignet",
+    "build:react-secure-biometric": "cd ../mosip-bio-device && npm run build-storybook -- -o ../storybook-example/storybook-static/react-sbi-integrator",
+    "build:sign-in-with-esignet": "cd ../sign-in-with-esignet && npm run build-storybook -- -o ../storybook-example/storybook-static/sign-in-with-esignet",
+    "build:secure-biometric": "cd ../secure-biometric-interface-integrator && npm run build-storybook -- -o ../storybook-example/storybook-static/secure-biometric-interface-integrator",
+    "build:children": "npm run build:secure-biometric && npm run build:sign-in-with-esignet && npm run build:react-secure-biometric && npm run build:react-sign-in-with-esignet",
+    "build:all": "npm run build && npm run build:children",
     "deploy": "gh-pages -d storybook-static",
-    "publish": "npm run build && npm run deploy"
+    "sbook:react-sign-in-with-esignet": "cd ../react-sign-in-with-esignet && npm run storybook -- --port 6003 --no-open",
+    "sbook:react-secure-biometric": "cd ../mosip-bio-device && npm run storybook -- --port 6004 --no-open",
+    "sbook:sign-in-with-esignet": "cd ../sign-in-with-esignet && npm run storybook -- --port 6001 --no-open",
+    "sbook:secure-biometric": "cd ../secure-biometric-interface-integrator && npm run storybook -- --port 6002 --no-open",
+    "sbook-children": "npx concurrently \"npm run sbook:sign-in-with-esignet\"  \"npm run sbook:secure-biometric\" \"npm run sbook:react-sign-in-with-esignet\"  \"npm run sbook:react-secure-biometric\"",
+    "sbook-parent": "npx wait-on http://localhost:6001 && npx wait-on http://localhost:6002 && npx wait-on http://localhost:6003 && npx wait-on http://localhost:6004 && npm run storybook",
+    "sbook": "npx concurrently \"npm run sbook-children\" \"npm run sbook-parent\"",
+    "publish": "npm run build:all && npm run deploy"
   },
   "author": "",
   "license": "MPL-2.0",
